# Conflicts:
#	COMG_DonovanSim/Source/SP2_Shorthand.cpp
#	COMG_DonovanSim/Source/Ship.cpp

+++ COMG_DonovanSim/Source/ShipParts.cpp
@@ -52,7 +52,7 @@ string ShipParts::getName()
     returns a value that is used to increase the ship's stats
 */
 /******************************************************************************/
-int ShipParts::partsEffect()
+float ShipParts::partsEffect()
 {
     return 0;
 }
